+++ spec/integration/resource/api500/c7000/ethernet_network/update_spec.rb
@@ -1,0 +1,7 @@
+require 'spec_helper'
+
+RSpec.describe OneviewSDK::API500::C7000::EthernetNetwork, integration: true, type: UPDATE do
+  include_examples 'EthernetNetworkUpdateExample', 'integration api500 context' do
+    let(:current_client) { $client_500 }
+  end
+end