+++ spec/spec_helper.rb
@@ -81,6 +81,7 @@ require_relative 'shared_context'
 require_relative 'support/fake_response'
 require_relative 'integration/sequence_and_naming'
 require_relative 'integration/sequence_and_naming_i3s'
+Dir["#{File.dirname(__FILE__)}/integration/shared_examples/**/*.rb"].each { |file| require file }
 
 RSpec.configure do |config|
   # Sort integration and system tests